Allow building gtsam with ccache or other compilers
If you build your workspace with a non-default compiler, gtsam still gets built with the system default compiler. This change allows passing the chosen compiler to the external project.

In my case, I want to use ccache, which speeds up subsequent clean builds of gtsam_catkin to seconds.

Maybe also `CMAKE_CXX_FLAGS` could be passed to gtsam, but I'm not sure if that would be 100% safe.